+++ tests/Type/GeneralTypeTest.php
@@ -30,6 +30,9 @@ class GeneralTypeTest extends TypeInferenceTestCase
         yield from $this->gatherAssertTypes(__DIR__.'/data/abstract-manager.php');
         yield from $this->gatherAssertTypes(__DIR__.'/data/facades.php');
         yield from $this->gatherAssertTypes(__DIR__.'/data/where-relation.php');
+        yield from $this->gatherAssertTypes(__DIR__.'/data/validator.php');
+        yield from $this->gatherAssertTypes(__DIR__.'/data/form-request.php');
+        yield from $this->gatherAssertTypes(__DIR__.'/data/database-transaction.php');
     }
 
     /**